Adapt it to ink#next with actual React
