+++ .gitignore
@@ -4,3 +4,4 @@
 /dist
 /tmp
 node_modules
+build/**
