+++ package.json
@@ -19,11 +19,18 @@
     "lodash": "^4.17.11",
     "log4js": "^3.0.6",
     "mustache": "^3.0.0",
+    "prop-types": "^15.6.2",
+    "react": "^16.6.3",
+    "react-redux": "^5.1.1",
     "recursive-copy": "^2.0.9",
     "redux": "^4.0.1",
     "underscore.string": "^3.3.5"
   },
   "devDependencies": {
+    "@babel/cli": "^7.1.5",
+    "@babel/core": "^7.1.6",
+    "@babel/plugin-transform-modules-commonjs": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
     "@oclif/dev-cli": "^1",
     "eslint": "^5.5",
     "eslint-config-oclif": "^3.1",
@@ -48,7 +55,7 @@
   "license": "MPL-2.0",
   "main": "src/index.js",
   "oclif": {
-    "commands": "./src/commands",
+    "commands": "./build/commands",
     "bin": "terrastack",
     "plugins": [
       "@oclif/plugin-help"
@@ -59,8 +66,17 @@
       }
     }
   },
+  "babel": {
+    "plugins": [
+      "@babel/plugin-transform-modules-commonjs"
+    ],
+    "presets": [
+      "@babel/preset-react"
+    ]
+  },
   "repository": "terrastackio/terrastack-cli",
   "scripts": {
+    "build": "babel src --out-dir build",
     "postpack": "rm -f oclif.manifest.json npm-shrinkwrap.json",
     "prepack": "oclif-dev manifest && oclif-dev readme && npm shrinkwrap",
     "test": "jest",
